refactored bbs2gh to split command/handlers
